+++ packages/ui5-test-writer/templates/v4/1.84/integration/ObjectPageJourney.ts
@@ -46,11 +46,15 @@ function journey() {
     opaTest("Navigate to <%- name%>ObjectPage", function (Given: Given, When: When, Then: Then) {
         Given.iStartMyApp();
 <% if(navigationParents.parentLRName) { -%>
+<% const parentTableId = navigationParents.parentLRViewKey ? '"' + navigationParents.parentLRViewKey + '"' : '""'; -%>
 <% if (!hideFilterBar) { -%>
         When.onThe<%- navigationParents.parentLRName%>Generated.onFilterBar().iExecuteSearch();
 <% } -%>
-        Then.onThe<%- navigationParents.parentLRName%>Generated.onTable(<%- navigationParents.parentLRTableIdentifier ? '"' + navigationParents.parentLRTableIdentifier + '"' : '""' %>).iCheckRows();
-        When.onThe<%- navigationParents.parentLRName%>Generated.onTable(<%- navigationParents.parentLRTableIdentifier ? '"' + navigationParents.parentLRTableIdentifier + '"' : '""' %>).iPressRow(0);
+<% if (navigationParents.parentLRViewKey && !navigationParents.parentLRViewIsDefault) { -%>
+        When.onThe<%- navigationParents.parentLRName%>Generated.iGoToView({ key: "<%- navigationParents.parentLRViewKey %>" });
+<% } -%>
+        Then.onThe<%- navigationParents.parentLRName%>Generated.onTable(<%- parentTableId %>).iCheckRows();
+        When.onThe<%- navigationParents.parentLRName%>Generated.onTable(<%- parentTableId %>).iPressRow(0);
 <% } -%>
 <% navigationParents.parentOPs.forEach(function(parent) { %>
         Then.onThe<%- parent.name %>Generated.iSeeThisPage();
@@ -80,12 +84,6 @@ function journey() {
     });
 <% } -%>
 
-<% if (headerTitle) { -%>
-    opaTest("Check header title of the Object Page", function (_Given: Given, _When: When, Then: Then) {
-        Then.onThe<%- name%>Generated.onHeader().iCheckTitlePath(<%- JSON.stringify(headerTitle) %>);
-    });
-
-<% } -%>
 <% if (headerSections?.length > 0) { -%>
     opaTest("Check header facets of the Object Page", function (_Given: Given, _When: When, Then: Then) {
 <% headerSections.forEach(function(section) { -%>
@@ -115,7 +113,7 @@ function journey() {
 <% } -%>
 <% if (bodySections?.length > 0) { -%>
 <% bodySections.forEach(function(section) { -%>
-    opaTest("Check the <%- section.id %> section of the Object Page", function (_Given: Given, <% if (bodySections?.length > 1) { %>When: When<% } else { %>_When: When<% } %>, Then: Then) {
+    opaTest("Check the <%- section.id %> section of the Object Page", function (_Given: Given, <% if (bodySections.length > 1) { %>When: When<% } else { %>_When: When<% } %>, Then: Then) {
 <% if (bodySections.length > 1) { -%>
         When.onThe<%- name%>Generated.iPressSectionIconTabFilterButton("<%- section.id %>");
 <% } -%>
@@ -153,8 +151,10 @@ function journey() {
 <% } -%>
 <% if (section?.subSections?.length > 0) { -%>
 <% section.subSections.forEach(function(subSection) { -%>
+<% if (section.subSections.length > 1) { -%>
         //When.onThe<%- name%>Generated.iGoToSection({ section: "<%- section.id %>", subSection: "<%- subSection.id %>" });
         Then.onThe<%- name%>Generated.iCheckSubSection({ section: "<%- subSection.id %>" }, {});
+<% } -%>
 <% if (subSection.fields && subSection.fields.length > 0) { -%>
 <% subSection.fields.forEach(function(field) { -%>
         Then.onThe<%- name%>Generated.onForm({ section: "<%- subSection.id %>" } as unknown as FormIdentifier).iCheckField({ property: "<%- field.property %>"<% if (field.connectedFields) { %>, connectedFields: "<%- field.connectedFields %>"<% } %><% if (field.fieldGroup) { %>, fieldGroup: "<%- field.fieldGroup %>"<% } %> });