+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 1.0.46
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.2.7 → 1.2.8
+- @sap-ux/inquirer-common 1.0.32 → 1.0.33
+- @sap-ux/flp-config-inquirer 1.0.45 → 1.0.46
+- @sap-ux/adp-tooling 1.0.45 → 1.0.46
+
+## 1.0.45
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/axios-extension 2.0.8 → 2.0.9
+- @sap-ux/inquirer-common 1.0.31 → 1.0.32
+- @sap-ux/project-access 2.1.10 → 2.1.11
+- @sap-ux/adp-tooling 1.0.44 → 1.0.45
+- @sap-ux/fiori-generator-shared 1.2.6 → 1.2.7
+- @sap-ux/flp-config-inquirer 1.0.44 → 1.0.45
+- @sap-ux/system-access 1.0.10 → 1.0.11
+
+## 1.0.44
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.43 → 1.0.44
+- @sap-ux/flp-config-inquirer 1.0.43 → 1.0.44
+
+## 1.0.43
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.42 → 1.0.43
+- @sap-ux/flp-config-inquirer 1.0.42 → 1.0.43
+
 ## 1.0.42
 
 ### Patch Changes