+++ packages/backend-proxy-middleware-cf/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux/backend-proxy-middleware-cf
 
+## 0.3.46
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.45 → 1.0.46
+
+## 0.3.45
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.44 → 1.0.45
+
+## 0.3.44
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.43 → 1.0.44
+
+## 0.3.43
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.42 → 1.0.43
+
 ## 0.3.42
 
 ### Patch Changes